Refactor analysis task
